# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "961 Apart Docs",
        "description": "Description to Api",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v1/building/all": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Get all",
                "operationId": "a7b60f546ffe40dd728fc743fc8c4f95",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": [
                                        {
                                            "id": 1,
                                            "name": "Town house",
                                            "square_from": "150",
                                            "square_to": "300",
                                            "number_floors": 12,
                                            "number_apartments": 140,
                                            "number_parking_spaces": 200,
                                            "is_mock": false,
                                            "photo": "buildings/T1rx1kz7khqGlNz2C0EHu3hMs3XOurCMfPzOcOYs.jpg"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/building/{id}/find": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Retrieve Building by Id",
                "operationId": "a97151b51cb2f363cb812e362bb85b70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Building id",
                        "required": true
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "?include=address,locations,layouts,price",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": {
                                        "id": 14,
                                        "name": "Town house",
                                        "square_from": "150",
                                        "square_to": "300",
                                        "number_floors": 12,
                                        "number_apartments": 140,
                                        "number_parking_spaces": 200,
                                        "is_mock": false,
                                        "photo": "buildings/JpbCqIijlhiQxTVhKf3NPnrdQqO460v3treGyUZz.jpg",
                                        "address": {
                                            "id": 5,
                                            "building_id": 14,
                                            "address": "rteet",
                                            "longitude": "69.12",
                                            "latitude": "42.11"
                                        },
                                        "locations": [
                                            {
                                                "id": 1,
                                                "building_id": 14,
                                                "name": "Дендро парк в близи",
                                                "description": "Самый большой парк"
                                            }
                                        ],
                                        "layouts": [
                                            {
                                                "id": 4,
                                                "building_id": 14,
                                                "photo": "layouts/GMexOjnX8DzaThxnBtZ9FxvX9QuLyfhBS1gbhT1M.jpg",
                                                "square_from": "40",
                                                "square_to": "110",
                                                "rooms": 4
                                            }
                                        ],
                                        "prices": {
                                            "id": 4,
                                            "building_id": 14,
                                            "price_from": "100",
                                            "price_to": "1000",
                                            "discount": "35",
                                            "installment": "0"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
...
}